Revert "Merge branch 'master' into master-mine"
This reverts commit 74b69f5325e0a7a89bcb326b5e32386a77334fdb, reversing
changes made to 4a911af51d6a2fffe794805af99042f87640fcac.

+++ UndertaleModLib/Models/UndertaleFunction.cs
@@ -52,7 +52,7 @@ public class UndertaleFunction : UndertaleNamedResource, UndertaleInstruction.Re
                 writer.Write(addr);
         }
         else
-            writer.Write(-1);
+            writer.Write((int)-1);
     }
 
     /// <inheritdoc />
